+++ packages/plugin-packages/mashroom-portal-demo-alternative-theme/src/backend/plugin/dev_server.ts
@@ -41,6 +41,7 @@ app.get('/', (req: Request, res: Response) => {
         apiBasePath: '/api',
         resourcesBasePath: '/resources',
         lastThemeReloadTs: Date.now(),
+        themeVersionHash: 'aaaaa',
         site: {
             siteId: 'test',
             title: 'Test Site',